pyproject.toml
[tool.poetry]
name = "nonebot-plugin-gshisbanner"
version = "0.4"
description = "Nonebot2查询原神历史卡池小插件"
authors = ["forchannot <yy320206@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_gshisbanner"}]

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-rc.2"
nonebot-adapter-onebot = "^2.1"
httpx = ">=0.20.0, <1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
